[TwigBundle] Move the {% render ... %} node to a dedicated extension + Remove service container
The purpose of the TemplatingExtension is ambigous. This first step move the actions logical
to a dedicated extension and reduce the dependency to the HTTP Kernel.
