+++ src/Symfony/Bundle/TwigBundle/Resources/config/twig.xml
@@ -43,6 +43,11 @@
             <argument type="service" id="service_container" />
         </service>
 
+        <service id="twig.extension.actions" class="Symfony\Bundle\TwigBundle\Extension\ActionsExtension" public="false">
+            <tag name="twig.extension" />
+            <argument type="service" id="http_kernel" />
+        </service>
+
         <service id="twig.extension.routing" class="Symfony\Bridge\Twig\Extension\RoutingExtension" public="false">
             <tag name="twig.extension" />
             <argument type="service" id="router" />
